+++ src/main/kotlin/com/machiav3lli/fdroid/ui/compose/pages/app_detail/components/LinkItem.kt
@@ -17,13 +17,9 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.machiav3lli.fdroid.R
 import com.machiav3lli.fdroid.entity.LinkType
-import com.machiav3lli.fdroid.ui.compose.theme.AppTheme
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
@@ -59,7 +55,7 @@ fun LinkItem(
                 maxLines = 1,
                 style = MaterialTheme.typography.titleMedium
             )
-            if(linkType.link!=null){
+            if (linkType.link != null) {
                 Text(
                     text = linkType.link.toString(),
                     style = MaterialTheme.typography.bodySmall,
@@ -69,18 +65,5 @@ fun LinkItem(
                 )
             }
         }
-    }
-}
-
-@Preview
-@Composable
-fun LinkItemPreview() {
-    AppTheme(blackTheme = false) {
-        LinkItem(
-            linkType = LinkType(
-                R.drawable.ic_email,
-                stringResource(id = R.string.author_email),
-                Uri.parse("neostore@neoapps.com"))
-        )
     }
 }
