+++ src/main/kotlin/com/machiav3lli/fdroid/ui/fragments/AppSheetX.kt
@@ -178,7 +178,7 @@ class AppSheetX() : FullscreenBottomSheetDialogFragment(), Callbacks {
         }
         viewModel.downloadState.value = state
         viewModel.updateActions()
-        if (downloadState is DownloadService.State.Success && !rootInstallerEnabled) { // && isResumed
+        if (downloadState is DownloadService.State.Success && !rootInstallerEnabled) { // && isResumed TODO unite root and normal install calls
             withContext(Dispatchers.Default) {
                 AppInstaller.getInstance(context)?.defaultInstaller?.install(downloadState.release.cacheFileName)
             }
